+++ bulk.py
@@ -7,14 +7,26 @@ class BulkExportSingleCommand(sublime_plugin.WindowCommand):
     def __init__(self, *args, **kwargs):
         super(BulkExportSingleCommand, self).__init__(*args, **kwargs)
 
-    def run(self):
+    def run(self, export_soql=False):
+        self.export_soql = export_soql
         sobjects_describe = processor.populate_sobjects_describe()
         self.sobjects = sorted(sobjects_describe.keys())
         self.window.show_quick_panel(self.sobjects, self.on_done)
 
     def on_done(self, index):
         if index == -1: return
-        processor.handle_backup_sobject_thread(self.sobjects[index])
+        self.chosen_sobject = self.sobjects[index]
+        if not self.export_soql:
+            processor.handle_backup_sobject_thread(self.chosen_sobject)
+        else:
+            self.window.show_input_panel("Input SOQL: ", "", self.on_input, None, None)
+
+    def on_input(self, soql):
+        if not soql or not soql.upper().startswith("SELECT"):
+            sublime.error_message("Invalid SOQL")
+            return
+
+        processor.handle_backup_sobject_thread(self.chosen_sobject, soql)
 
 class BulkExportAllCommand(sublime_plugin.WindowCommand):
     def __init__(self, *args, **kwargs):